requirements.txt
fastapi==0.77.1
python-dotenv==0.20.0

SQLAlchemy~=1.4.36